+++ setup.py
@@ -48,13 +48,10 @@ setup(
     
     # Old Spanish profiles
     spain_dcat=ckanext.dcat.profiles:SpanishDCATProfile
-    spain_dcat_ap=ckanext.dcat.profiles:SpanishDCATAPProfile
 
     ## New Spanish profiles
     es_dcat=ckanext.dcat.profiles:SpanishDCATProfile
     es_dcat_ap_2=ckanext.dcat.profiles:SpanishDCATAP2Profile
-    #es_geodcat_ap_3=ckanext.dcat.profiles:SpanishGeoDCATAP3Profile
-    #es_dcat_ap_3=ckanext.dcat.profiles:SpanishDCATAP3Profile
 
     [babel.extractors]
     ckan = ckan.lib.extract:extract_ckan
